renamed generic_utilities.h to generic_utilities.hpp
