+++ applications/GeoMechanicsApplication/custom_workflows/dgeosettlement.cpp
@@ -31,7 +31,7 @@
 #include "custom_processes/deactivate_conditions_on_inactive_elements_process.hpp"
 #include "custom_processes/find_neighbour_elements_of_conditions_process.h"
 #include "custom_processes/geo_extrapolate_integration_point_values_to_nodes_process.h"
-#include "custom_utilities/generic_utilities.h"
+#include "custom_utilities/generic_utilities.hpp"
 #include "custom_utilities/input_utility.h"
 #include "custom_utilities/process_info_parser.h"
 #include "custom_utilities/process_utilities.h"
