+++ applications/GeoMechanicsApplication/tests/cpp_tests/custom_utilities/test_generic_utilities.cpp
@@ -13,7 +13,7 @@
 //                   Gennady Markelov
 //
 
-#include "custom_utilities/generic_utilities.h"
+#include "custom_utilities/generic_utilities.hpp"
 #include "tests/cpp_tests/geo_mechanics_fast_suite.h"
 #include <boost/numeric/ublas/assignment.hpp>
